Code cleanup
The JNI-Layer now actually compiles, and the Makefile creates a libjtox.so
with 'make all'. The current makefile works exclusively on Linux, a portable
build system will be introduced later. In order for 'make all' to work, it assumes
that libtoxcore is installed on your system.
To try if linking the library worked, execute the JNITest.java file located in the test folder.
If Java complains about missing libraries, add '-Djava.library.path=jni:/usr/lib:/usr/local/lib'
to the VM arguments in your Run configuration, so Java knows where to search for libraries