Stop exposing the pointer to the internal data structures
After long and thorugh considerations, I decided that this was the only way to provide a sane API. If you
want to pass around JTox instances on Android, you will now have to create a Wrapper class that extends JTox
and implements Parcelable. I will create a guide on how to do this later. This way, the API is a lot cleaner,
and a lot of potential pitfalls for API users have been eliminated.