Refactor and set_statusmessage wrapper
Refactor some of the error codes to fit the entire API, instead of just adding of friends.
Also implement the wrapper for setting our own status message. If the message is longer than or equal to TOX_MAX_STATUSMESSAGE_LENGTH bytes, an exception will be thrown, in order to prevent truncation by the core. The same
applies to setname, but the value that must not be reached is TOX_MAX_NICKNAME_LENGTH.