Friendlist API:
Create a new friendlist API that handles a friend list that is synchronized with core. All callbacks
automatically update the list, and it is restored on loading. Read-Interaction with friends should now take
place through the friend list, instead of the JTox instance.

In addition, the test client is removed for now. The test directory is reserved for actual (unit)-tests. A reference client will be created in https://github.com/Tox/Macchiatox.